refactor: festival 이 stage 나 ticket 을 모르도록 변경
