+++ backend/src/main/java/com/festago/festival/application/FestivalService.java
@@ -1,7 +1,5 @@
 package com.festago.festival.application;
 
-import static java.util.Comparator.comparing;
-
 import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.NotFoundException;
@@ -15,8 +13,6 @@ import com.festago.festival.repository.FestivalFilter;
 import com.festago.festival.repository.FestivalRepository;
 import com.festago.school.domain.School;
 import com.festago.school.repository.SchoolRepository;
-import com.festago.stage.domain.Stage;
-import com.festago.stage.repository.StageRepository;
 import java.time.Clock;
 import java.time.LocalDate;
 import java.util.List;
@@ -31,7 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class FestivalService {
 
     private final FestivalRepository festivalRepository;
-    private final StageRepository stageRepository;
+    private final FestivalStageService festivalStageService;
     private final SchoolRepository schoolRepository;
     private final Clock clock;
 
@@ -57,11 +53,7 @@ public class FestivalService {
 
     @Transactional(readOnly = true)
     public DetailFestivalResponse findDetail(Long festivalId) {
-        Festival festival = findFestival(festivalId);
-        List<Stage> stages = stageRepository.findAllDetailByFestivalId(festivalId).stream()
-            .sorted(comparing(Stage::getStartTime))
-            .toList();
-        return DetailFestivalResponse.of(festival, stages);
+        return festivalStageService.findDetail(festivalId);
     }
 
     private Festival findFestival(Long festivalId) {
