+++ backend/src/main/java/com/festago/festival/application/FestivalStageService.java
@@ -1,0 +1,10 @@
+package com.festago.festival.application;
+
+import com.festago.festival.dto.DetailFestivalResponse;
+import org.springframework.stereotype.Component;
+
+@Component
+public interface FestivalStageService {
+
+    DetailFestivalResponse findDetail(Long festivalId);
+}
