+++ backend/src/main/java/com/festago/festival/dto/DetailFestivalResponse.java
@@ -1,8 +1,5 @@
 package com.festago.festival.dto;
 
-import com.festago.festival.domain.Festival;
-import com.festago.stage.domain.Stage;
-import com.festago.stage.dto.DetailStageResponse;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -15,18 +12,4 @@ public record DetailFestivalResponse(
     String thumbnail,
     List<DetailStageResponse> stages) {
 
-    public static DetailFestivalResponse of(Festival festival, List<Stage> stages) {
-        List<DetailStageResponse> stageResponses = stages.stream()
-            .map(DetailStageResponse::from)
-            .toList();
-        return new DetailFestivalResponse(
-            festival.getId(),
-            festival.getSchool().getId(),
-            festival.getName(),
-            festival.getStartDate(),
-            festival.getEndDate(),
-            festival.getThumbnail(),
-            stageResponses
-        );
-    }
 }
