+++ backend/src/main/java/com/festago/festival/dto/DetailTicketResponse.java
@@ -1,0 +1,9 @@
+package com.festago.festival.dto;
+
+public record DetailTicketResponse(
+    Long id,
+    String ticketType,
+    Integer totalAmount,
+    Integer remainAmount) {
+
+}
