+++ backend/src/test/java/com/festago/festival/application/FestivalServiceTest.java
@@ -13,13 +13,13 @@ import com.festago.common.exception.BadRequestException;
 import com.festago.common.exception.NotFoundException;
 import com.festago.festival.domain.Festival;
 import com.festago.festival.dto.DetailFestivalResponse;
+import com.festago.festival.dto.DetailStageResponse;
 import com.festago.festival.dto.FestivalCreateRequest;
 import com.festago.festival.dto.FestivalResponse;
 import com.festago.festival.repository.FestivalRepository;
 import com.festago.school.domain.School;
 import com.festago.school.repository.SchoolRepository;
 import com.festago.stage.domain.Stage;
-import com.festago.stage.dto.DetailStageResponse;
 import com.festago.stage.repository.StageRepository;
 import com.festago.support.FestivalFixture;
 import com.festago.support.SchoolFixture;
