+++ backend/src/test/java/com/festago/festival/application/integration/FestivalServiceIntegrationTest.java
@@ -5,13 +5,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.festago.festival.application.FestivalService;
 import com.festago.festival.domain.Festival;
 import com.festago.festival.dto.DetailFestivalResponse;
+import com.festago.festival.dto.DetailStageResponse;
+import com.festago.festival.dto.DetailTicketResponse;
 import com.festago.festival.dto.FestivalCreateRequest;
 import com.festago.festival.dto.FestivalResponse;
 import com.festago.festival.repository.FestivalRepository;
 import com.festago.school.domain.School;
 import com.festago.school.repository.SchoolRepository;
 import com.festago.stage.domain.Stage;
-import com.festago.stage.dto.DetailStageResponse;
 import com.festago.stage.repository.StageRepository;
 import com.festago.support.ApplicationIntegrationTest;
 import com.festago.support.FestivalFixture;
@@ -20,7 +21,6 @@ import com.festago.support.StageFixture;
 import com.festago.support.TicketFixture;
 import com.festago.ticket.domain.Ticket;
 import com.festago.ticket.domain.TicketType;
-import com.festago.ticket.dto.DetailTicketResponse;
 import com.festago.ticket.repository.TicketRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
